solved several issues inside  * imports causes bad behaviour for classes for nspawn docker etc,  * separated also general class to file  * fixed issues with escaping commands  * removed reprecaded symlinks
