--- examples/testing-module/moduleframework
@@ -1,1 +1,0 @@
-../../moduleframework/
