--- examples/testing-module/modulelint
@@ -1,1 +1,0 @@
-../../tools/modulelint
